alsaloop: Preserve playback/capture buffer sync on xrun
When using the shared buffer, the alsaloop playback and capture
pointers may currently get out of sync because buf_pos is handled
wrong. When injecting silence samples, the buf_pos pointer should
be decremented (modulo buf_size) and the buf_count should then be
incremented with the same amount. This way the buffer would just
rewind after the next writeit() call.

This patch fixes the operations on play->buf_pos in order to keep
it in sync with the capture side.

Signed-off-by: Dragos Tarcatu <dragos_tarcatu@mentor.com>